+++ packages/client/package.json
@@ -45,13 +45,13 @@
   "license": "MIT",
   "devDependencies": {
     "@accounts/client": "0.0.11",
-    "@accounts/common": "0.1.0",
+    "@accounts/common": "canary",
     "coveralls": "^2.11.14",
     "node-fetch": "^1.7.3"
   },
   "peerDependencies": {
     "@accounts/client": "^0.0.15",
-    "@accounts/common": "^0.1.0"
+    "@accounts/common": "canary"
   },
   "dependencies": {
     "lodash": "^4.17.4"
